[.github\workflows\cmake.yml] Set `VCPKG_TARGET_TRIPLET`; [fswatch\src\fswatch.cpp] Use windows optarg; use secure variants of `gmtime` & `localtime`; [fswatch\src\CMakeLists.txt] Add getopt dir/lib on MSVC; [fswatch\src\windows] getopt port for MSVC; [libfswatch] MSVC variants

+++ libfswatch/src/libfswatch/c++/poll_monitor.cpp
@@ -228,7 +228,7 @@ namespace fsw
       FSW_ELOG(_("Done scanning.\n"));
 
 #ifdef _MSC_VER
-      Sleep( 1000 * (
+      Sleep( 1000L * static_cast<DWORD>(
 #else
      sleep( (
 #endif
